chore: Split round syscall implementation to multiple source files
