--- core/src/syscall/precompiles/blake2s/round.rs
@@ -1,1111 +1,0 @@
-use crate::bytes::event::ByteRecord;
-use crate::memory::{MemoryCols, MemoryReadCols, MemoryWriteCols};
-use crate::operations::{Add4Operation, FixedRotateRightOperation, XorOperation};
-use crate::runtime::{MemoryReadRecord, MemoryWriteRecord};
-use crate::stark::SphinxAirBuilder;
-use crate::utils::pad_rows;
-use crate::{
-    air::{EventLens, MachineAir, WithEvents},
-    runtime::{ExecutionRecord, Program, Syscall, SyscallCode, SyscallContext},
-};
-use core::borrow::{Borrow, BorrowMut};
-use p3_air::{Air, AirBuilder, BaseAir};
-use p3_field::{AbstractField, PrimeField32};
-use p3_matrix::dense::RowMajorMatrix;
-use p3_matrix::Matrix;
-use serde::Deserialize;
-use serde::Serialize;
-use sphinx_derive::AlignedBorrow;
-use std::mem::size_of;
-
-const R_1: u32 = 16;
-const R_2: u32 = 12;
-const R_3: u32 = 8;
-const R_4: u32 = 7;
-
-#[derive(Default)]
-pub struct Blake2sRoundChip;
-
-impl Blake2sRoundChip {
-    pub fn new() -> Self {
-        Blake2sRoundChip
-    }
-    pub fn constrain_shuffled_indices<AB: SphinxAirBuilder>(
-        &self,
-        builder: &mut AB,
-        indices: &[AB::Var],
-        is_real: AB::Var,
-    ) {
-        for index in 0..4 {
-            builder
-                .when(is_real)
-                .assert_eq(indices[index], AB::F::from_canonical_usize(0));
-        }
-        for index in 4..indices.len() {
-            builder
-                .when(is_real)
-                .assert_eq(indices[index], AB::F::from_canonical_usize(1));
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Blake2sRoundEvent {
-    pub lookup_id: usize,
-    pub clk: u32,
-    pub shard: u32,
-    pub channel: u32,
-    pub a_ptr: u32,
-    pub b_ptr: u32,
-
-    pub a_reads_writes: Vec<MemoryWriteRecord>,
-    pub b_reads: Vec<MemoryReadRecord>,
-}
-
-impl Syscall for Blake2sRoundChip {
-    fn execute(&self, ctx: &mut SyscallContext<'_, '_>, arg1: u32, arg2: u32) -> Option<u32> {
-        let clk_init = ctx.clk;
-        let shard = ctx.current_shard();
-        let lookup_id = ctx.syscall_lookup_id;
-        let channel = ctx.current_channel();
-
-        let a_ptr = arg1;
-        let b_ptr = arg2;
-
-        // a: v[0] || v[1] || v[2] || v[3] || v[4] || v[5] || v[6] || v[7] || v[8] || v[9] || v[10] || v[11] || v[12] || v[13] || v[14] || v[15] ||
-        let mut a = ctx.slice_unsafe(a_ptr, 16);
-        let mut a_clone = a.clone();
-
-        // b: m1[0] || m1[1] || m1[2] || m1[3] || || m2[0] || m2[1] || m2[2] || m2[3] ||
-        //    m3[0] || m3[1] || m3[2] || m3[3] || m4[0] || m4[1] || m4[2] || mr[3] ||
-        //    0 || 0 || 0 || 0 || 0 || 0 || 0 || 0
-        let (b_reads, b) = ctx.mr_slice(b_ptr, 24);
-        assert_eq!(b[16], 0u32);
-        assert_eq!(b[17], 0u32);
-        assert_eq!(b[18], 0u32);
-        assert_eq!(b[19], 0u32);
-        assert_eq!(b[20], 0u32);
-        assert_eq!(b[21], 0u32);
-        assert_eq!(b[22], 0u32);
-        assert_eq!(b[23], 0u32);
-
-        // 1x (m1, R1, R2)
-        // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m1)
-        for ((v0, v1), m) in a[0..4]
-            .iter_mut()
-            .zip(a_clone[4..8].iter())
-            .zip(b[0..4].iter())
-        {
-            *v0 = (*v0).wrapping_add(*v1).wrapping_add(*m);
-        }
-        a_clone = a.clone();
-
-        // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-        for (v3, v0) in a[12..16].iter_mut().zip(a_clone[0..4].iter()) {
-            *v3 = (*v3 ^ *v0).rotate_right(R_1);
-        }
-        a_clone = a.clone();
-
-        // v[2] = v[2].wrapping_add(v[3]);
-        for (v2, v3) in a[8..12].iter_mut().zip(a_clone[12..16].iter()) {
-            *v2 = (*v2).wrapping_add(*v3);
-        }
-        a_clone = a.clone();
-
-        // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-        for (v1, v2) in a[4..8].iter_mut().zip(a_clone[8..12].iter()) {
-            *v1 = (*v1 ^ *v2).rotate_right(R_2);
-        }
-
-        // 2x (m2, R3, R4)
-        let mut a = a.clone(); // a after 1x quarter_round
-        let mut a_clone = a.clone();
-
-        // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m2)
-        for ((v0, v1), m) in a[0..4]
-            .iter_mut()
-            .zip(a_clone[4..8].iter())
-            .zip(b[4..8].iter())
-        {
-            *v0 = (*v0).wrapping_add(*v1).wrapping_add(*m);
-        }
-        a_clone = a.clone();
-
-        // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-        for (v3, v0) in a[12..16].iter_mut().zip(a_clone[0..4].iter()) {
-            *v3 = (*v3 ^ *v0).rotate_right(R_3);
-        }
-        a_clone = a.clone();
-
-        // v[2] = v[2].wrapping_add(v[3]);
-        for (v2, v3) in a[8..12].iter_mut().zip(a_clone[12..16].iter()) {
-            *v2 = (*v2).wrapping_add(*v3);
-        }
-        a_clone = a.clone();
-
-        // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-        for (v1, v2) in a[4..8].iter_mut().zip(a_clone[8..12].iter()) {
-            *v1 = (*v1 ^ *v2).rotate_right(R_4);
-        }
-
-        // shuffle
-        // v[1]
-        a[4..8].swap(0, 3);
-        a[4..8].swap(0, 1);
-        a[4..8].swap(1, 2);
-
-        // v[2]
-        a[8..12].swap(0, 2);
-        a[8..12].swap(1, 3);
-
-        // v[3]
-        a[12..16].swap(2, 3);
-        a[12..16].swap(1, 2);
-        a[12..16].swap(0, 1);
-
-        let mut a = a.clone(); // a after 2x quarter_round
-        let mut a_clone = a.clone();
-
-        // 3x (m3, R1, R2)
-        // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m3)
-        for ((v0, v1), m) in a[0..4]
-            .iter_mut()
-            .zip(a_clone[4..8].iter())
-            .zip(b[8..12].iter())
-        {
-            *v0 = (*v0).wrapping_add(*v1).wrapping_add(*m);
-        }
-        a_clone = a.clone();
-
-        // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-        for (v3, v0) in a[12..16].iter_mut().zip(a_clone[0..4].iter()) {
-            *v3 = (*v3 ^ *v0).rotate_right(R_1);
-        }
-        a_clone = a.clone();
-
-        // v[2] = v[2].wrapping_add(v[3]);
-        for (v2, v3) in a[8..12].iter_mut().zip(a_clone[12..16].iter()) {
-            *v2 = (*v2).wrapping_add(*v3);
-        }
-        a_clone = a.clone();
-
-        // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-        for (v1, v2) in a[4..8].iter_mut().zip(a_clone[8..12].iter()) {
-            *v1 = (*v1 ^ *v2).rotate_right(R_2);
-        }
-
-        // 4x (m4, R3, R4)
-        let mut a = a.clone(); // a after 3x quarter_round
-        let mut a_clone = a.clone();
-
-        // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m2)
-        for ((v0, v1), m) in a[0..4]
-            .iter_mut()
-            .zip(a_clone[4..8].iter())
-            .zip(b[12..16].iter())
-        {
-            *v0 = (*v0).wrapping_add(*v1).wrapping_add(*m);
-        }
-        a_clone = a.clone();
-
-        // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-        for (v3, v0) in a[12..16].iter_mut().zip(a_clone[0..4].iter()) {
-            *v3 = (*v3 ^ *v0).rotate_right(R_3);
-        }
-        a_clone = a.clone();
-
-        // v[2] = v[2].wrapping_add(v[3]);
-        for (v2, v3) in a[8..12].iter_mut().zip(a_clone[12..16].iter()) {
-            *v2 = (*v2).wrapping_add(*v3);
-        }
-        a_clone = a.clone();
-
-        // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-        for (v1, v2) in a[4..8].iter_mut().zip(a_clone[8..12].iter()) {
-            *v1 = (*v1 ^ *v2).rotate_right(R_4);
-        }
-
-        // unshuffle
-        // v[1]
-        a[4..8].swap(2, 3);
-        a[4..8].swap(1, 2);
-        a[4..8].swap(0, 1);
-
-        // v[2]
-        a[8..12].swap(0, 2);
-        a[8..12].swap(1, 3);
-
-        // v[3]
-        a[12..16].swap(0, 3);
-        a[12..16].swap(0, 1);
-        a[12..16].swap(1, 2);
-
-        ctx.clk += 1;
-        // Write rotate_right to a_ptr.
-        let a_reads_writes = ctx.mw_slice(a_ptr, a.as_slice());
-
-        ctx.record_mut()
-            .blake2s_round_events
-            .push(Blake2sRoundEvent {
-                lookup_id,
-                clk: clk_init,
-                shard,
-                channel,
-                a_ptr,
-                b_ptr,
-                a_reads_writes,
-                b_reads,
-            });
-
-        None
-    }
-
-    fn num_extra_cycles(&self) -> u32 {
-        1
-    }
-}
-
-#[derive(Debug, Clone, AlignedBorrow)]
-#[repr(C)]
-struct Blake2sRoundCols<T> {
-    pub clk: T,
-    pub shard: T,
-    pub channel: T,
-    pub nonce: T,
-    pub is_real: T,
-
-    pub a_ptr: T,
-    pub b_ptr: T,
-
-    pub shuffled_indices: [T; 16],
-
-    pub a: [MemoryWriteCols<T>; 16],
-    pub b: [MemoryReadCols<T>; 24],
-
-    pub add: [Add4Operation<T>; 32],
-    pub xor: [XorOperation<T>; 32],
-    pub rotate_right: [FixedRotateRightOperation<T>; 32],
-}
-
-impl<T: PrimeField32> BaseAir<T> for Blake2sRoundChip {
-    fn width(&self) -> usize {
-        size_of::<Blake2sRoundCols<u8>>()
-    }
-}
-
-impl<'a> WithEvents<'a> for Blake2sRoundChip {
-    type Events = &'a [Blake2sRoundEvent];
-}
-
-impl<F: PrimeField32> MachineAir<F> for Blake2sRoundChip {
-    type Record = ExecutionRecord;
-    type Program = Program;
-
-    fn name(&self) -> String {
-        "Blake2sRoundChip".to_string()
-    }
-
-    fn generate_trace<EL: EventLens<Self>>(
-        &self,
-        input: &EL,
-        output: &mut Self::Record,
-    ) -> RowMajorMatrix<F> {
-        let mut rows = vec![];
-        let width = <Blake2sRoundChip as BaseAir<F>>::width(self);
-        let mut new_byte_lookup_events = Vec::new();
-        for event in input.events() {
-            let shard = event.shard;
-            let mut row = vec![F::zero(); width];
-            let cols: &mut Blake2sRoundCols<F> = row.as_mut_slice().borrow_mut();
-
-            cols.clk = F::from_canonical_u32(event.clk);
-            cols.is_real = F::one();
-            cols.shard = F::from_canonical_u32(event.shard);
-            cols.channel = F::from_canonical_u32(event.channel);
-            cols.a_ptr = F::from_canonical_u32(event.a_ptr);
-            cols.b_ptr = F::from_canonical_u32(event.b_ptr);
-
-            // populate all v, m
-            for i in 0..16usize {
-                cols.a[i].populate(
-                    event.channel,
-                    event.a_reads_writes[i],
-                    &mut new_byte_lookup_events,
-                );
-                cols.b[i].populate(event.channel, event.b_reads[i], &mut new_byte_lookup_events);
-            }
-
-            // populate zeroes
-            for i in 16..24usize {
-                cols.b[i].populate(event.channel, event.b_reads[i], &mut new_byte_lookup_events);
-            }
-
-            let mut v0_outer = vec![0u32; 4];
-            let mut v1_outer = vec![0u32; 4];
-            let mut v2_outer = vec![0u32; 4];
-            let mut v3_outer = vec![0u32; 4];
-
-            // 1x (m1, R1, R2)
-            for i in 0..4usize {
-                let v0 = event.a_reads_writes[i].prev_value;
-                let v1 = event.a_reads_writes[i + 4].prev_value;
-                let v2 = event.a_reads_writes[i + 8].prev_value;
-                let v3 = event.a_reads_writes[i + 12].prev_value;
-                let m1 = event.b_reads[i].value;
-                let zero1 = event.b_reads[i + 16].value;
-                let zero2 = event.b_reads[i + 20].value;
-                assert_eq!(zero1, 0);
-                assert_eq!(zero2, 0);
-
-                // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-                let v0_new = cols.add[i].populate(output, shard, event.channel, v0, v1, m1, zero1);
-                assert_eq!(v0 + v1 + m1 + zero1, v0_new);
-
-                // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-                let temp = cols.xor[i].populate(output, shard, event.channel, v3, v0_new);
-                let v3_new =
-                    cols.rotate_right[i].populate(output, shard, event.channel, temp, R_1 as usize);
-                assert_eq!((v3 ^ v0_new).rotate_right(R_1), v3_new);
-
-                // v[2] = v[2].wrapping_add(v[3]);
-                let v2_new = cols.add[i + 4].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    v2,
-                    v3_new,
-                    zero1,
-                    zero2,
-                );
-                assert_eq!(v2 + v3_new + zero1 + zero2, v2_new);
-
-                // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-                let temp = cols.xor[i + 4].populate(output, shard, event.channel, v1, v2_new);
-                let v1_new = cols.rotate_right[i + 4].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_2 as usize,
-                );
-                assert_eq!((v1 ^ v2_new).rotate_right(R_2), v1_new);
-
-                v0_outer[i] = v0_new;
-                v1_outer[i] = v1_new;
-                v2_outer[i] = v2_new;
-                v3_outer[i] = v3_new;
-            }
-
-            // 2x (m2, R3, R4)
-            for i in 0..4usize {
-                let v0 = v0_outer[i];
-                let v1 = v1_outer[i];
-                let v2 = v2_outer[i];
-                let v3 = v3_outer[i];
-                let m2 = event.b_reads[i + 4].value;
-                let zero1 = event.b_reads[i + 16].value;
-                let zero2 = event.b_reads[i + 20].value;
-                assert_eq!(zero1, 0);
-                assert_eq!(zero2, 0);
-
-                // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m2)
-                let v0_new =
-                    cols.add[i + 8].populate(output, shard, event.channel, v0, v1, m2, zero1);
-                assert_eq!(v0 + v1 + m2 + zero1, v0_new);
-
-                // v[3] = (v[3] ^ v[0]).rotate_right_const(rd); (R3)
-                //cols.shuffled_indices[i + 12] = F::from_canonical_u32(1);
-                let temp = cols.xor[i + 8].populate(output, shard, event.channel, v3, v0_new);
-                //let v3_new = cols.rotate_right[v3_shuffle_lookup[i] + 8].populate(
-                let v3_new = cols.rotate_right[i + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_3 as usize,
-                );
-                assert_eq!((v3 ^ v0_new).rotate_right(R_3), v3_new);
-
-                // v[2] = v[2].wrapping_add(v[3]);
-                //cols.shuffled_indices[i + 8] = F::from_canonical_u32(1);
-                //let v2_new = cols.add[v2_shuffle_lookup[i] + 4 + 8].populate(
-                let v2_new = cols.add[i + 4 + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    v2,
-                    v3_new,
-                    zero1,
-                    zero2,
-                );
-                assert_eq!(v2 + v3_new + zero1 + zero2, v2_new);
-
-                // v[1] = (v[1] ^ v[2]).rotate_right_const(rb); (R4)
-                //cols.shuffled_indices[i + 4] = F::from_canonical_u32(1);
-                let temp = cols.xor[i + 4 + 8].populate(output, shard, event.channel, v1, v2_new);
-                //let v1_new = cols.rotate_right[v1_shuffle_lookup[i] + 4 + 8].populate(
-                let v1_new = cols.rotate_right[i + 4 + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_4 as usize,
-                );
-                assert_eq!((v1 ^ v2_new).rotate_right(R_4), v1_new);
-
-                v0_outer[i] = v0_new;
-                v1_outer[i] = v1_new;
-                v2_outer[i] = v2_new;
-                v3_outer[i] = v3_new;
-            }
-
-            // shuffle
-            // v[1]
-            v1_outer.swap(0, 3);
-            v1_outer.swap(0, 1);
-            v1_outer.swap(1, 2);
-
-            // v[2]
-            v2_outer.swap(0, 2);
-            v2_outer.swap(1, 3);
-
-            // v[3]
-            v3_outer.swap(2, 3);
-            v3_outer.swap(1, 2);
-            v3_outer.swap(0, 1);
-
-            // 3x (m3, R1, R2)
-            for i in 0..4usize {
-                cols.shuffled_indices[i + 4] = F::from_canonical_u32(1);
-                cols.shuffled_indices[i + 8] = F::from_canonical_u32(1);
-                cols.shuffled_indices[i + 12] = F::from_canonical_u32(1);
-
-                let v0 = v0_outer[i];
-                let v1 = v1_outer[i];
-                let v2 = v2_outer[i];
-                let v3 = v3_outer[i];
-                let m3 = event.b_reads[i + 8].value;
-                let zero1 = event.b_reads[i + 16].value;
-                let zero2 = event.b_reads[i + 20].value;
-                assert_eq!(zero1, 0);
-                assert_eq!(zero2, 0);
-
-                // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-                let v0_new =
-                    cols.add[i + 16].populate(output, shard, event.channel, v0, v1, m3, zero1);
-                assert_eq!(v0 + v1 + m3 + zero1, v0_new);
-
-                // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-                let temp = cols.xor[i + 16].populate(output, shard, event.channel, v3, v0_new);
-                let v3_new = cols.rotate_right[i + 16].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_1 as usize,
-                );
-                assert_eq!((v3 ^ v0_new).rotate_right(R_1), v3_new);
-
-                // v[2] = v[2].wrapping_add(v[3]);
-                let v2_new = cols.add[i + 16 + 4].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    v2,
-                    v3_new,
-                    zero1,
-                    zero2,
-                );
-                assert_eq!(v2 + v3_new + zero1 + zero2, v2_new);
-
-                // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-                let temp = cols.xor[i + 16 + 4].populate(output, shard, event.channel, v1, v2_new);
-                let v1_new = cols.rotate_right[i + 16 + 4].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_2 as usize,
-                );
-                assert_eq!((v1 ^ v2_new).rotate_right(R_2), v1_new);
-
-                v0_outer[i] = v0_new;
-                v1_outer[i] = v1_new;
-                v2_outer[i] = v2_new;
-                v3_outer[i] = v3_new;
-            }
-
-            // 4x (m4, R3, R4)
-            for i in 0..4usize {
-                let v0 = v0_outer[i];
-                let v1 = v1_outer[i];
-                let v2 = v2_outer[i];
-                let v3 = v3_outer[i];
-                let m4 = event.b_reads[i + 12].value;
-                let zero1 = event.b_reads[i + 16].value;
-                let zero2 = event.b_reads[i + 20].value;
-                assert_eq!(zero1, 0);
-                assert_eq!(zero2, 0);
-
-                // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le()); (m2)
-                let v0_new =
-                    cols.add[i + 16 + 8].populate(output, shard, event.channel, v0, v1, m4, zero1);
-                assert_eq!(v0 + v1 + m4 + zero1, v0_new);
-
-                // v[3] = (v[3] ^ v[0]).rotate_right_const(rd); (R3)
-                //cols.unshuffled_indices[i + 12] = F::from_canonical_u32(1);
-                let temp = cols.xor[i + 16 + 8].populate(output, shard, event.channel, v3, v0_new);
-                //let v3_new = cols.rotate_right[v3_unshuffle_lookup[i] + 16 + 8].populate(
-                let v3_new = cols.rotate_right[i + 16 + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_3 as usize,
-                );
-                assert_eq!((v3 ^ v0_new).rotate_right(R_3), v3_new);
-
-                // v[2] = v[2].wrapping_add(v[3]);
-                //cols.unshuffled_indices[i + 8] = F::from_canonical_u32(1);
-                //let v2_new = cols.add[v2_unshuffle_lookup[i] + 16 + 4 + 8].populate(
-                let v2_new = cols.add[i + 16 + 4 + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    v2,
-                    v3_new,
-                    zero1,
-                    zero2,
-                );
-                assert_eq!(v2 + v3_new + zero1 + zero2, v2_new);
-
-                // v[1] = (v[1] ^ v[2]).rotate_right_const(rb); (R4)
-                //cols.unshuffled_indices[i + 4] = F::from_canonical_u32(1);
-                let temp =
-                    cols.xor[i + 16 + 4 + 8].populate(output, shard, event.channel, v1, v2_new);
-                //let v1_new = cols.rotate_right[v1_unshuffle_lookup[i] + 16 + 4 + 8].populate(
-                let v1_new = cols.rotate_right[i + 16 + 4 + 8].populate(
-                    output,
-                    shard,
-                    event.channel,
-                    temp,
-                    R_4 as usize,
-                );
-                assert_eq!((v1 ^ v2_new).rotate_right(R_4), v1_new);
-
-                //v0_outer[i] = v0_new;
-                //v1_outer[i] = v1_new;
-                //v2_outer[i] = v2_new;
-                //v3_outer[i] = v3_new;
-            }
-
-            /*
-            // unshuffle
-            // v[1]
-            v1_outer.swap(2, 3);
-            v1_outer.swap(1, 2);
-            v1_outer.swap(0, 1);
-
-            // v[2]
-            v2_outer.swap(0, 2);
-            v2_outer.swap(1, 3);
-
-            // v[3]
-            v3_outer.swap(0, 3);
-            v3_outer.swap(0, 1);
-            v3_outer.swap(1, 2);*/
-
-            //println!("4x traces: {:02x?}", [v0_outer, v1_outer, v2_outer, v3_outer].concat().to_vec());
-
-            rows.push(row);
-        }
-
-        output.add_byte_lookup_events(new_byte_lookup_events);
-
-        pad_rows(&mut rows, || {
-            let row = vec![F::zero(); width];
-
-            row
-        });
-
-        let mut trace =
-            RowMajorMatrix::<F>::new(rows.into_iter().flatten().collect::<Vec<_>>(), width);
-
-        // Write the nonces to the trace.
-        for i in 0..trace.height() {
-            let cols: &mut Blake2sRoundCols<F> =
-                trace.values[i * width..(i + 1) * width].borrow_mut();
-            cols.nonce = F::from_canonical_usize(i);
-        }
-
-        trace
-    }
-
-    fn included(&self, shard: &Self::Record) -> bool {
-        !shard.blake2s_round_events.is_empty()
-    }
-}
-
-impl<AB: SphinxAirBuilder> Air<AB> for Blake2sRoundChip
-where
-    AB::F: PrimeField32,
-{
-    fn eval(&self, builder: &mut AB) {
-        let main = builder.main();
-        let local = main.row_slice(0);
-        let next = main.row_slice(1);
-        let local: &Blake2sRoundCols<AB::Var> = (*local).borrow();
-        let next: &Blake2sRoundCols<AB::Var> = (*next).borrow();
-
-        // Constrain the incrementing nonce.
-        builder.when_first_row().assert_zero(local.nonce);
-        builder
-            .when_transition()
-            .assert_eq(local.nonce + AB::Expr::one(), next.nonce);
-
-        for i in 0..16usize {
-            // Eval a
-            builder.eval_memory_access(
-                local.shard,
-                local.channel,
-                local.clk + AB::F::from_canonical_u32(1), // We eval 'a' pointer access at clk+1 since 'a', 'b' could be the same,
-                local.a_ptr + AB::F::from_canonical_u32((i as u32) * 4),
-                &local.a[i],
-                local.is_real,
-            );
-
-            // Eval b
-            builder.eval_memory_access(
-                local.shard,
-                local.channel,
-                local.clk,
-                local.b_ptr + AB::F::from_canonical_u32((i as u32) * 4),
-                &local.b[i],
-                local.is_real,
-            );
-        }
-
-        // zeroes
-        for i in 16..24usize {
-            // Eval b
-            builder.eval_memory_access(
-                local.shard,
-                local.channel,
-                local.clk,
-                local.b_ptr + AB::F::from_canonical_u32((i as u32) * 4),
-                &local.b[i],
-                local.is_real,
-            );
-        }
-
-        let v1_shuffle_lookup = vec![1, 2, 3, 0];
-        let v2_shuffle_lookup = vec![2, 3, 0, 1];
-        let v3_shuffle_lookup = vec![3, 0, 1, 2];
-
-        for i in 0..4usize {
-            // 1x
-
-            // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-            Add4Operation::<AB::F>::eval(
-                builder,
-                *local.a[i].prev_value(),     // v0
-                *local.a[i + 4].prev_value(), // v1
-                *local.b[i].value(),          // m1
-                *local.b[i + 20].value(),     // zero1
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i],
-            );
-
-            // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                *local.a[i + 12].prev_value(),
-                local.add[i].value,
-                local.xor[i],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i].value,
-                R_1 as usize,
-                local.rotate_right[i],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // v[2] = v[2].wrapping_add(v[3]);
-            Add4Operation::<AB::F>::eval(
-                builder,
-                *local.a[i + 8].prev_value(), // v2
-                local.rotate_right[i].value,  // v3 from previous operation
-                *local.b[i + 16].value(),     // zero1
-                *local.b[i + 20].value(),     // zero2
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 4],
-            );
-
-            // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                *local.a[i + 4].prev_value(),
-                local.add[i + 4].value,
-                local.xor[i + 4],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 4].value,
-                R_2 as usize,
-                local.rotate_right[i + 4],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // 2x
-            // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[i].value,              // v0 after 1x
-                local.rotate_right[i + 4].value, // v1 after 1x
-                *local.b[i + 4].value(),         // m2
-                *local.b[i + 16].value(),        // zero1
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 8],
-            );
-
-            // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[i].value, // v3 after 1x
-                local.add[i + 8].value,      // v0 after 1x
-                local.xor[i + 8],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 8].value,
-                R_3 as usize,
-                local.rotate_right[i + 8],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // v[2] = v[2].wrapping_add(v[3]);
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[i + 4].value,          // v2 after 1x
-                local.rotate_right[i + 8].value, // v3 after previous operation
-                *local.b[i + 16].value(),        // zero1
-                *local.b[i + 20].value(),        // zero2
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 12],
-            );
-
-            // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[i + 4].value, // v1 after 1x
-                local.add[i + 12].value,         // v2 after previous operation
-                local.xor[i + 12],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 12].value,
-                R_4 as usize,
-                local.rotate_right[i + 12],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-        }
-
-        self.constrain_shuffled_indices(builder, &local.shuffled_indices, local.is_real);
-
-        for i in 0..4usize {
-            // 3x
-            // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[i + 8].value, // v0 after 2x
-                local.rotate_right[v1_shuffle_lookup[i] + 12].value, // v1 after 2x
-                *local.b[i + 8].value(), // m3
-                *local.b[i + 16].value(), // zero1
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 16],
-            );
-
-            // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[v3_shuffle_lookup[i] + 8].value, // v3 after 2x
-                local.add[i + 16].value,                            // v0 after previous operation
-                local.xor[i + 16],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 16].value,
-                R_1 as usize,
-                local.rotate_right[i + 16],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // v[2] = v[2].wrapping_add(v[3]);
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[v2_shuffle_lookup[i] + 12].value, // v2 after 2x
-                local.rotate_right[i + 16].value,           // v3 after previous operation
-                *local.b[i + 16].value(),                   // zero1
-                *local.b[i + 20].value(),                   // zero2
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 20],
-            );
-
-            // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[v1_shuffle_lookup[i] + 12].value, // v1 after 2x
-                local.add[i + 20].value,                             // v2 after previous operation
-                local.xor[i + 20],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 20].value,
-                R_2 as usize,
-                local.rotate_right[i + 20],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // 4x
-            // v[0] = v[0].wrapping_add(v[1]).wrapping_add(m.from_le());
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[i + 16].value,          // v0 after 3x
-                local.rotate_right[i + 20].value, // v1 after 3x
-                *local.b[i + 12].value(),         // m4
-                *local.b[i + 16].value(),         // zero1
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 24],
-            );
-
-            // v[3] = (v[3] ^ v[0]).rotate_right_const(rd);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[i + 16].value, // v3 after 3x
-                local.add[i + 24].value,          // v0 after previous operation
-                local.xor[i + 24],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 24].value,
-                R_3 as usize,
-                local.rotate_right[i + 24],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // v[2] = v[2].wrapping_add(v[3]);
-            Add4Operation::<AB::F>::eval(
-                builder,
-                local.add[i + 20].value,          // v2 after 3x
-                local.rotate_right[i + 24].value, // v3 after previous operation
-                *local.b[i + 16].value(),         // zero1
-                *local.b[i + 20].value(),         // zero2
-                local.shard,
-                local.channel,
-                local.is_real,
-                local.add[i + 28],
-            );
-
-            // v[1] = (v[1] ^ v[2]).rotate_right_const(rb);
-            // Eval XOR
-            XorOperation::<AB::F>::eval(
-                builder,
-                local.rotate_right[i + 20].value, // v1 after 3x
-                local.add[i + 28].value,          // v2 after previous operation
-                local.xor[i + 28],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-
-            // Eval RotateRight
-            FixedRotateRightOperation::<AB::F>::eval(
-                builder,
-                local.xor[i + 28].value,
-                R_4 as usize,
-                local.rotate_right[i + 28],
-                local.shard,
-                &local.channel,
-                local.is_real,
-            );
-        }
-
-        /*self.constrain_shuffled_unshuffled_indices(
-            builder,
-            &local.unshuffled_indices,
-            local.is_real,
-        );*/
-
-        builder.receive_syscall(
-            local.shard,
-            local.channel,
-            local.clk,
-            local.nonce,
-            AB::F::from_canonical_u32(SyscallCode::BLAKE_2S_ROUND.syscall_id()),
-            local.a_ptr,
-            local.b_ptr,
-            local.is_real,
-        )
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::runtime::{Instruction, Opcode, SyscallCode};
-    use crate::syscall::precompiles::blake2s::blake2s_round;
-    use crate::utils::tests::BLAKE2S_ROUND_ELF;
-    use crate::utils::{run_test, run_test_with_memory_inspection, setup_logger};
-    use crate::Program;
-    use rand::Rng;
-
-    fn risc_v_program(a_ptr: u32, b_ptr: u32, a: [u32; 16], b: [u32; 24]) -> Program {
-        let mut instructions = vec![];
-        // memory write a
-        for (index, word) in a.into_iter().enumerate() {
-            instructions.push(Instruction::new(Opcode::ADD, 29, 0, word, false, true));
-            instructions.push(Instruction::new(
-                Opcode::ADD,
-                30,
-                0,
-                a_ptr + (index * 4) as u32,
-                false,
-                true,
-            ));
-            instructions.push(Instruction::new(Opcode::SW, 29, 30, 0, false, true));
-        }
-
-        // memory write b
-        for (index, word) in b.into_iter().enumerate() {
-            instructions.push(Instruction::new(Opcode::ADD, 29, 0, word, false, true));
-            instructions.push(Instruction::new(
-                Opcode::ADD,
-                30,
-                0,
-                b_ptr + (index * 4) as u32,
-                false,
-                true,
-            ));
-            instructions.push(Instruction::new(Opcode::SW, 29, 30, 0, false, true));
-        }
-
-        // Syscall invocation
-        instructions.push(Instruction::new(
-            Opcode::ADD,
-            5,
-            0,
-            SyscallCode::BLAKE_2S_ROUND as u32,
-            false,
-            true,
-        ));
-        instructions.push(Instruction::new(Opcode::ADD, 10, 0, a_ptr, false, true));
-        instructions.push(Instruction::new(Opcode::ADD, 11, 0, b_ptr, false, true));
-        instructions.push(Instruction::new(Opcode::ECALL, 5, 10, 11, false, false));
-        Program::new(instructions, 0, 0)
-    }
-
-    #[test]
-    fn test_blake2s_round_precompile() {
-        setup_logger();
-
-        let a_ptr = 100100100;
-        let b_ptr = 200200200;
-
-        let a = rand::thread_rng().gen::<[u32; 16]>();
-        let mut a_clone = a.clone();
-        let mut b = rand::thread_rng().gen::<[u32; 24]>();
-        for item in b[0..24].iter_mut() {
-            *item = 0;
-        }
-
-        blake2s_round(&mut a_clone, &b);
-
-        let program = risc_v_program(a_ptr, b_ptr, a, b);
-
-        let (_, memory) = run_test_with_memory_inspection(program);
-        let mut result = vec![];
-        // result is 4 words, written to a_ptr
-        for i in 0..16 {
-            result.push(memory.get(&(a_ptr + i * 4)).unwrap().value);
-        }
-
-        assert_eq!(result, a_clone.to_vec());
-    }
-
-    #[test]
-    fn test_blake2s_round_program() {
-        setup_logger();
-        let program = Program::from(BLAKE2S_ROUND_ELF);
-        run_test(program).unwrap();
-    }
-}
